- Added A_ClearReFire code pointer for weapons. Preferably A_WeaponReady should   reset this counter but that can't be done due to unwanted side effects with   existing weapons. - Changed the 'scale' variable in CVAR(turbo) to double because the calculations   depended on the current floating point precision setting and only worked properly   when set to 'precise' in VC++.
SVN r1057 (trunk)

+++ src/g_game.cpp
@@ -217,7 +217,7 @@ CUSTOM_CVAR (Float, turbo, 100.f, 0)
 	}
 	else
 	{
-		float scale = self * 0.01f;
+		double scale = self * 0.01;
 
 		forwardmove[0] = (int)(normforwardmove[0]*scale);
 		forwardmove[1] = (int)(normforwardmove[1]*scale);
